add metaDelegateCall_EIP1271 function to Account.sol
this adds support for smart contracts as the account proxyOwner, as long as the proxyOwner implements EIP-1271 `isValidSignature`. The primary use case for adding this is to support GnosisSafe contracts as Brink signers.

The metaDelegateCall function exists to provide support for EOA signers. metaDelegateCall_EIP1271 won't work for EOA signers, it can only be used if the proxyOwner is a smart contract.