Change the proxy owner address to a constant
The primary motivation for this change is to make proxy owner addresses immutable, and therefore impossible for a "verifier" contract to overwrite. Previously, the owner address was stored in the first storage slot, so it was prone to being overwritten by a verifier contract using internal storage variables in solidity. We could also store the owner at a random storage pointer, but this would still allow a malicious verifier to overwrite it with another address.

This change adds a constant placeholder address `0xfefeFEFeFEFEFEFEFeFefefefefeFEfEfefefEfe` to the Proxy.sol bytecode. This is replaced with a real owner address by AccountFactory.sol when a proxy is deployed. On delegatecall, the proxy owner address can be read by the Account.sol implementation using `extcodecopy`. Unfortunately `codecopy` does not run within the context of the delegatecall executor contract - when delegatecall is executed from Proxy.sol to Account.sol, `codecopy` reads from the Account.sol bytecode instead of the Proxy.sol bytecode. The `extcodecopy` technique still uses less gas than our previous builds.